storage: move disk unlocking to the installation method screen
This change adjusts the design to allow users to unlock their LUKS devices before
selecting an installation scenario.

The motivation for this adjustment is to enable the 'home reuse' scenario to identify
it's availability, by viewing the encrypted disk contents.
