+++ src/components/storage/EncryptedDevices.jsx
@@ -21,8 +21,8 @@ import React, { useState } from "react";
 import {
     ActionList,
     ActionListItem,
+    Alert,
     Button,
-    Divider,
     Flex,
     FlexItem,
     Form,
@@ -30,8 +30,7 @@ import {
     InputGroup,
     InputGroupItem,
     Modal,
-    Stack,
-    StackItem, TextInput,
+    TextInput,
 } from "@patternfly/react-core";
 import { EyeIcon, EyeSlashIcon, LockIcon } from "@patternfly/react-icons";
 
@@ -41,7 +40,6 @@ import {
 
 import { getDevicesAction } from "../../actions/storage-actions.js";
 
-import { EmptyStatePanel } from "cockpit-components-empty-state.jsx";
 import { FormHelper } from "cockpit-components-form-helper.jsx";
 import { InlineNotification } from "cockpit-components-inline-notification.jsx";
 
@@ -60,37 +58,26 @@ const LuksDevices = ({ id, lockedLUKSDevices }) => {
     );
 };
 
-export const EncryptedDevices = ({ dispatch, idPrefix, lockedLUKSDevices, setSkipUnlock }) => {
+export const EncryptedDevices = ({ dispatch, idPrefix, lockedLUKSDevices }) => {
     const [showUnlockDialog, setShowUnlockDialog] = useState(false);
     return (
         <>
-            <Divider />
-            <EmptyStatePanel
-              title={_("Encrypted devices are locked")}
-              paragraph={
-                  <Stack hasGutter className={idPrefix + "-empty-state-body"}>
-                      <StackItem>{_("Devices should be unlocked before assigning mount points.")}</StackItem>
-                      <StackItem>
-                          <LuksDevices lockedLUKSDevices={lockedLUKSDevices} />
-                      </StackItem>
-                  </Stack>
-              }
-              secondary={
+            <Alert
+              isInline
+              title={_("Destination is encrypted")}
+              variant="warning"
+              actionLinks={
                   <ActionList>
                       <ActionListItem>
                           <Button id={idPrefix + "-unlock-devices-btn"} variant="primary" onClick={() => setShowUnlockDialog(true)}>
-                              {_("Unlock devices")}
-                          </Button>
-                      </ActionListItem>
-                      <ActionListItem>
-                          <Button variant="secondary" onClick={() => setSkipUnlock(true)}>
-                              {_("Skip")}
+                              {_("Unlock")}
                           </Button>
                       </ActionListItem>
                   </ActionList>
               }
-            />
-            <Divider />
+            >
+                {_("Unlock LUKS-encrypted partitions to keep existing data and show more installation methods.")}
+            </Alert>
             {showUnlockDialog &&
             <UnlockDialog
               dispatch={dispatch}
